+++ utils/commonerrors/errors.go
@@ -37,6 +37,7 @@ var (
 	ErrForbidden          = errors.New("forbidden")
 	ErrCondition          = errors.New("failed condition")
 	ErrEOF                = errors.New("end of file")
+	ErrMalicious          = errors.New("suspected malicious intent")
 )
 
 func Any(target error, err ...error) bool {
