+++ utils/git/limits.go
@@ -8,6 +8,8 @@ import (
 	"github.com/ARM-software/golang-utils/utils/config"
 )
 
+// TODO: Make these use apply
+
 type noLimits struct {
 }
